# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Adare General Hospital Platform API",
    "version": "1.0.0",
    "description": "REST API for the Adare General Hospital digital platform (Hawassa, Sidama Region, Ethiopia). All responses use the envelope { ok, data, message } or { ok:false, code, error }. Protected endpoints require Authorization: Bearer <access_token>; refresh tokens rotate via HTTP-only cookie on /api/auth."
  },
  "servers": [{ "url": "/api" }],
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } },
    "schemas": {
...
      "PaymentCreate": {
        "type": "object",
        "required": ["payer_name", "phone", "amount", "method"],
        "properties": {
          "payer_name": { "type": "string" }, "phone": { "type": "string" },
          "amount": { "type": "number" }, "method": { "enum": ["telebirr", "bank_transfer", "card", "cash", "cbhi", "other"] },
          "provider_ref": { "type": "string" }, "appointment_ref": { "type": "string" }
        }
      }
    }
  },
  "paths": {
...
    "/patients/me": {
      "get": { "security": [{ "bearerAuth": [] }], "summary": "Patient portal: own profile + appointments + payments + notifications", "responses": { "200": { "description": "Profile" } } },
      "patch": { "security": [{ "bearerAuth": [] }], "summary": "Update own profile", "responses": { "200": { "description": "Updated" } } }
    },
    "/patients": {
      "post": { "security": [{ "bearerAuth": [] }], "summary": "Staff: register patient → AGH-PAT number", "responses": { "201": { "description": "Patient" } } },
      "get": { "security": [{ "bearerAuth": [] }], "summary": "Staff: search patients", "parameters": [{ "name": "q", "in": "query", "required": true }], "responses": { "200": { "description": "Matches" } } }
    },
    "/payments": {
      "post": { "summary": "Submit payment → AGH-PAY reference (PENDING)", "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaymentCreate" } } } }, "responses": { "201": { "description": "Payment" }, "409": { "description": "Duplicate provider ref" } } },
      "get": { "security": [{ "bearerAuth": [] }], "summary": "Finance/reception payment list", "responses": { "200": { "description": "Payments" } } }
    },
    "/payments/{reference}": { "patch": { "security": [{ "bearerAuth": [] }], "summary": "Finance: SUCCESSFUL | FAILED | CANCELLED | REFUNDED", "responses": { "200": { "description": "Updated" }, "409": { "description": "Invalid transition" } } } },
    "/payments/summary": { "get": { "security": [{ "bearerAuth": [] }], "summary": "Revenue KPIs", "responses": { "200": { "description": "Summary" } } } },
    "/payments/export/csv": { "get": { "security": [{ "bearerAuth": [] }], "summary": "CSV export (finance)", "responses": { "200": { "description": "CSV file" } } } },
...
  }
}